LayoutTests/fast/parser/template-multiple-child-text-nodes-expected.txt

This tests inserting multiple text tokens inside a template element. WebKit should create a single contiguous text node.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS templates[0].content.firstChild.data is "hello"
PASS templates[1].content.childNodes[1].data is "world"
PASS successfullyParsed is true

TEST COMPLETE
